broke utils.py into mutiple files
