+++ app.py
@@ -1,6 +1,7 @@
 from flask import Flask, request, jsonify
 import os
-from utils import extract_text, preprocess_text, rank_resume, analyze_pdf_formatting, check_consistency
+from utils.text_processing import extract_text, preprocess_text, rank_resume
+from utils.formatting import analyze_pdf_formatting, check_consistency
 
 app = Flask(__name__)
 
@@ -10,9 +11,9 @@ RESUME_DIR = "resume_samples"  # Directory to store resumes
 def analyze_resume():
     """
     Analyze a resume against a job description by performing:
-    1. Keyword and similarity analysis.
-    2. Overall formatting analysis.
-    3. Consistency checks (including headings and vertical spacing).
+      1. Keyword and similarity analysis.
+      2. Overall formatting analysis.
+      3. Consistency checks (headings and vertical spacing).
     Returns a combined JSON response with detailed feedback.
     """
     data = request.json
@@ -30,12 +31,10 @@ def analyze_resume():
     if not resume_text:
         return jsonify({"error": "Could not extract text from resume"}), 500
 
-    # Preprocess texts for keyword analysis
     resume_text_processed = preprocess_text(resume_text)
     job_desc_processed = preprocess_text(job_desc)
     keyword_result = rank_resume(resume_text_processed, job_desc_processed)
 
-    # Overall formatting analysis
     formatting_results = analyze_pdf_formatting(resume_path)
     formatting_messages = []
     if formatting_results["unique_font_names"] > 1:
@@ -47,7 +46,6 @@ def analyze_resume():
     if not formatting_messages:
         formatting_messages.append("Overall formatting looks good!")
     
-    # Detailed consistency checks (headings and vertical spacing)
     consistency_messages = check_consistency(resume_path)
     formatting_messages.extend(consistency_messages)
